+++ mod.ts
@@ -1,8 +1,8 @@
-import { CheerioAPI, load } from 'cheerio';
-import { ContentTabs, LemmaElement, Links, Selectors, zip } from './src/mod.ts';
+import { Cheerio, CheerioAPI, Element, load } from 'cheerio';
+import { ContentTabs, Links, Selectors, zip } from './src/mod.ts';
 
 namespace Dexonline {
-	export async function get(word: string): Promise<Array<LemmaElement> | undefined> {
+	export async function get(word: string): Promise<Array<Synthesis.Lemma> | undefined> {
 		const response = await fetch(Links.definition(word));
 		if (!response.ok) return undefined;
 
@@ -10,27 +10,50 @@ namespace Dexonline {
 		return parse(body);
 	}
 
-	export function parse(body: string): Array<LemmaElement> {
+	export function parse(body: string): Array<Synthesis.Lemma> {
 		const $ = load(body);
 
-		const entries = parseEntries($);
-
-		console.debug(entries);
+		const entries = Synthesis.parseEntries($);
 
 		return entries;
 	}
 
-	function parseEntries($: CheerioAPI): Array<LemmaElement> {
-		const synthesis = $(Selectors.contentTab(ContentTabs.Synthesis));
+	namespace Synthesis {
+		export interface Lemma extends Header {}
 
-		const headersBodies = zip(
-			synthesis.find(Selectors.contentTabs.synthesis.header).toArray(),
-			synthesis.find(Selectors.contentTabs.synthesis.body).toArray(),
-		);
+		interface Header {
+			type: string;
+			lemma: string;
+		}
 
-		return headersBodies.map(
-			([header, body]) => ({ header: $(header), body: $(body) }),
-		);
+		export function parseEntries($: CheerioAPI): Array<Lemma> {
+			const synthesis = $(Selectors.contentTab(ContentTabs.Synthesis));
+
+			const headerBodyTuples = zip(
+				synthesis.find(Selectors.contentTabs.synthesis.header.element).toArray(),
+				synthesis.find(Selectors.contentTabs.synthesis.body.element).toArray(),
+			);
+
+			return headerBodyTuples.map(
+				([headerElement, _bodyElement]) => {
+					const header = parseHeader($(headerElement));
+
+					return { ...header };
+				},
+			);
+		}
+
+		export function parseHeader(element: Cheerio<Element>): Header {
+			const typeElement = element.find(Selectors.contentTabs.synthesis.header.type);
+			const type = typeElement.text().trim().toLowerCase();
+
+			typeElement.remove();
+
+			const [singular, _plural] = <[string, string]> element.text().trim().split(', ');
+			const lemma = singular;
+
+			return { type, lemma };
+		}
 	}
 }