+++ src/constants.ts
@@ -8,8 +8,13 @@ namespace Selectors {
 	export const contentTab = (tab: ContentTabs) => `#tab_${tab}`;
 	export const contentTabs = {
 		synthesis: {
-			header: 'h3[class=tree-heading] > div',
-			body: 'div[class=tree-body]',
+			header: {
+        element: 'h3[class=tree-heading] > div',
+        type: 'span[class=tree-pos-info]',
+      },
+			body: {
+        element: 'div[class=tree-body]',
+      },
 		},
 	};
 }