+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 ## 1.0.0 (Unreleased)
+
+FEATURES:
+
+- Parsing entries in the 'Synthesis' tab:
+  - Headers:
+    - Lemma
+    - Type